Move the row translation utitilites into their own file
