+++ entrytool/management/commands/load_demographics.py
@@ -1,42 +1,11 @@
-import datetime
 import csv
 from django.core.management.base import BaseCommand
 from django.db import transaction
-from entrytool.models import PatientDetails
+from entrytool.management.commands.load_utils import (
+    translate_date, get_and_check, no_yes_unknown, get_or_create_ll
+)
+from entrytool.models import PatientDetails, Hospital
 from opal.models import Patient
-
-
-def translate_date(some_str):
-    """
-    We expect day/month/year
-
-    We don't use strptime because we don't have leading 0s
-    """
-    if not some_str:
-        return None
-    day, month, year = some_str.strip().split("/")
-    if int(year) > datetime.date.today().year:
-        year = "19{}".format(year)
-    else:
-        year = "20{}".format(year)
-
-    return datetime.date(int(year), int(month), int(day))
-
-
-def get_and_check(row_value, choices):
-    row_value = row_value.strip()
-    if not row_value:
-        return None
-    if row_value not in [i[0] for i in choices]:
-        raise ValueError("{} not in {}".format(row_value, choices))
-    return row_value
-
-
-def no_yes_unknown(row_value):
-    NO_YES_UNKNOWN = {
-        0: "No", 1: "Yes", 2: "Unknown"
-    }
-    return NO_YES_UNKNOWN.get(int(row_value))
 
 
 class Command(BaseCommand):
